refactor: refactor FCMservice for proper initialization and foreground notification handling
