+++ app/src/main/java/campus/tech/kakao/map/presentation/viewmodel/WelcomeViewModel.kt
@@ -1,19 +1,16 @@
 package campus.tech.kakao.map.presentation.viewmodel
 
-import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import campus.tech.kakao.map.data.remote.config.RemoteConfigManager
-import campus.tech.kakao.map.data.repository.WelcomeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 @HiltViewModel
 class WelcomeViewModel @Inject constructor(
     private val remoteConfigManager: RemoteConfigManager,
-    private val repository: WelcomeRepository
 ): ViewModel() {
     private val _serviceState = MutableLiveData<String>()
     val serviceState: LiveData<String> get() = _serviceState
@@ -43,13 +40,5 @@ class WelcomeViewModel @Inject constructor(
                 _serviceMessage.value = "Failed to load..."
             }
         }
-    }
-
-    fun createNotificationChannel(context: Context) {
-        repository.createNotificationChannel(context)
-    }
-
-    fun sendForegroundNotification(context: Context) {
-        repository.sendForegroundNotification(context)
     }
 }
